+++ src/rstr/rstr_p2.cpp
@@ -410,16 +410,11 @@ int32_t p2_Cstr2Cell(CSTR_line lin, CSTR_rast first, CSTR_rast last,
 		c1 = c2;
 
 		// Nick add
-		c2->dupstart = (int32_t) CSTR_GetDup(curr);
-		c2->dupend = (int32_t) CSTR_GetDupEnd(curr);
+		c2->dupstart = CSTR_GetDup(curr);
+		c2->dupend = CSTR_GetDupEnd(curr);
 
 		numCell++;
 	}
-
-	//c2=cell_l();
-	//c1->next=c1->nextl=c2;
-	//c2->prev=c2->prevl=c1;
-
 
 	return numCell;
 }
@@ -433,8 +428,6 @@ static void p2_CopyAttr2CSTR(CSTR_rast_attr *attr, cell *c) {
 	attr->r_row = c->r_row;
 	attr->r_col = c->r_col;
 
-	//attr->row=attr->r_row-(int16_t)((int32_t)nIncline*attr->r_col/2048);
-	//attr->col=attr->r_col+(int16_t)((int32_t)nIncline*attr->r_row/2048);
 	attr->row = c->row;
 	attr->col = c->col;
